+++ src/services/ISCClient.ts
@@ -8,7 +8,7 @@ import { SailPointISCAuthenticationProvider } from "./AuthenticationProvider";
 import { compareByName } from "../utils";
 import { DEFAULT_ACCOUNTS_QUERY_PARAMS } from "../models/Account";
 import { DEFAULT_ENTITLEMENTS_QUERY_PARAMS } from "../models/Entitlements";
-import { Configuration, IdentityProfilesApi, IdentityProfile, LifecycleState, LifecycleStatesApi, Paginator, ServiceDeskIntegrationApi, ServiceDeskIntegrationDto, Source, SourcesApi, TransformsApi, WorkflowsBetaApi, WorkflowExecutionBeta, ConnectorRuleManagementBetaApi, ConnectorRuleResponseBeta, ConnectorRuleValidationResponseBeta, AccountsApi, AccountsApiListAccountsRequest, Account, EntitlementsBetaApi, PublicIdentity, SPConfigBetaApi, SpConfigImportResultsBeta, SpConfigJobBeta, ImportOptionsBeta, SpConfigExportResultsBeta, ObjectExportImportOptionsBeta, TransformRead, GovernanceGroupsBetaApi, WorkgroupDtoBeta, AccessProfilesApiListAccessProfilesRequest, RolesApi, Role, RolesApiListRolesRequest, Search, SearchApi, IdentityDocument, SearchDocument, AccessProfileDocument, EntitlementDocument, RoleDocument, SourcesBetaApi, StatusResponseBeta, Schema, FormBeta, CustomFormsBetaApi, ExportFormDefinitionsByTenant200ResponseInnerBeta, FormDefinitionResponseBeta, NotificationsBetaApi, TemplateDtoBeta, SegmentsApi, Segment, SearchAttributeConfigurationBetaApi, SearchAttributeConfigBeta, IdentityAttributesBetaApi, IdentityAttributeBeta, PasswordManagementBetaApi, ConnectorRuleUpdateRequestBeta, IdentitiesBetaApi, IdentitiesBetaApiListIdentitiesRequest, IdentityBeta, IdentitySyncJobBeta, TaskResultResponseBeta, LoadEntitlementTaskBeta, TaskStatusBeta, EntitlementSourceResetBaseReferenceDtoBeta, TaskResultDtoBeta, ProvisioningPolicyDto, ImportFormDefinitionsRequestInnerBeta, ManagedClustersBetaApi, StandardLevelBeta, CertificationCampaignsV2025Api, CertificationsV2025Api, CertificationCampaignsV2025ApiMoveRequest, CertificationSummariesV2025Api, IdentityCertDecisionSummaryV2025, AccessReviewItemV2025, CertificationsV2025ApiReassignIdentityCertificationsRequest, CertificationsV2025ApiMakeIdentityDecisionRequest, IdentityCertificationDtoV2025, GetActiveCampaigns200ResponseInnerV2025, CertificationsV2025ApiSubmitReassignCertsAsyncRequest, WorkflowsApi, ExportPayloadBetaIncludeTypesBeta, SODPoliciesV2024Api, SodPolicyV2024, CertificationTask, AppsBetaApi, SourceAppBeta, ConfigurationHubV2024Api, BackupResponseV2024, IdentityProfilesV2025Api, IdentityPreviewResponseV2025, IdentityAttributeTransformV2025, SourcesV2025Api, TaskManagementV2025Api, AttributeDTO, RolesV2025Api, AccessProfilesV2025Api, JsonPatchOperationV2025, DimensionsV2025Api, RolesV2025ApiListRolesRequest, DimensionV2025, DimensionsV2025ApiListDimensionsRequest, PasswordConfigurationV2025Api, PasswordOrgConfigV2025, WorkflowsV2025Api, WorkflowV2025, WorkflowBodyV2025, WorkflowExecutionEventV2025, CreateWorkflowRequestV2025, WorkflowExecutionV2025, ConnectorRuleManagementV2025Api, ConnectorRuleResponseV2025, EntitlementsV2025Api, EntitlementsV2025ApiListEntitlementsRequest, EntitlementV2025, PublicIdentitiesV2025Api, PublicIdentitiesV2025ApiGetPublicIdentitiesRequest, RoleV2025, TransformsV2025Api, TransformV2025, TransformReadV2025, SearchV2025Api, IndexV2025, AccessProfileDocumentV2025, EntitlementDocumentV2025, RoleDocumentV2025, EventDocumentV2025, AccountActivityDocumentV2025, AccessProfileV2025 } from 'sailpoint-api-client';
+import { Configuration, IdentityProfilesApi, IdentityProfile, LifecycleState, LifecycleStatesApi, Paginator, ServiceDeskIntegrationApi, ServiceDeskIntegrationDto, Source, SourcesApi, TransformsApi, WorkflowsBetaApi, WorkflowExecutionBeta, ConnectorRuleManagementBetaApi, ConnectorRuleResponseBeta, ConnectorRuleValidationResponseBeta, AccountsApi, AccountsApiListAccountsRequest, Account, EntitlementsBetaApi, PublicIdentity, SPConfigBetaApi, SpConfigImportResultsBeta, SpConfigJobBeta, ImportOptionsBeta, SpConfigExportResultsBeta, ObjectExportImportOptionsBeta, TransformRead, GovernanceGroupsBetaApi, WorkgroupDtoBeta, AccessProfilesApiListAccessProfilesRequest, RolesApi, Role, RolesApiListRolesRequest, Search, SearchApi, IdentityDocument, SearchDocument, AccessProfileDocument, EntitlementDocument, RoleDocument, SourcesBetaApi, StatusResponseBeta, Schema, CustomFormsBetaApi, ExportFormDefinitionsByTenant200ResponseInnerBeta, FormDefinitionResponseV2025, CustomFormsV2025Api, CreateFormDefinitionRequestV2025, NotificationsBetaApi, TemplateDtoBeta, SegmentsApi, Segment, SearchAttributeConfigurationBetaApi, SearchAttributeConfigBeta, IdentityAttributesBetaApi, IdentityAttributeBeta, PasswordManagementBetaApi, ConnectorRuleUpdateRequestBeta, IdentitiesBetaApi, IdentitiesBetaApiListIdentitiesRequest, IdentityBeta, IdentitySyncJobBeta, TaskResultResponseBeta, LoadEntitlementTaskBeta, TaskStatusBeta, EntitlementSourceResetBaseReferenceDtoBeta, TaskResultDtoBeta, ProvisioningPolicyDto, ImportFormDefinitionsRequestInnerBeta, ManagedClustersBetaApi, StandardLevelBeta, CertificationCampaignsV2025Api, CertificationsV2025Api, CertificationCampaignsV2025ApiMoveRequest, CertificationSummariesV2025Api, IdentityCertDecisionSummaryV2025, AccessReviewItemV2025, CertificationsV2025ApiReassignIdentityCertificationsRequest, CertificationsV2025ApiMakeIdentityDecisionRequest, IdentityCertificationDtoV2025, GetActiveCampaigns200ResponseInnerV2025, CertificationsV2025ApiSubmitReassignCertsAsyncRequest, WorkflowsApi, ExportPayloadBetaIncludeTypesBeta, SODPoliciesV2024Api, SodPolicyV2024, CertificationTask, AppsBetaApi, SourceAppBeta, ConfigurationHubV2024Api, BackupResponseV2024, IdentityProfilesV2025Api, IdentityPreviewResponseV2025, IdentityAttributeTransformV2025, SourcesV2025Api, TaskManagementV2025Api, AttributeDTO, RolesV2025Api, AccessProfilesV2025Api, JsonPatchOperationV2025, DimensionsV2025Api, RolesV2025ApiListRolesRequest, DimensionV2025, DimensionsV2025ApiListDimensionsRequest, PasswordConfigurationV2025Api, PasswordOrgConfigV2025, WorkflowsV2025Api, WorkflowV2025, WorkflowBodyV2025, WorkflowExecutionEventV2025, CreateWorkflowRequestV2025, WorkflowExecutionV2025, ConnectorRuleManagementV2025Api, ConnectorRuleResponseV2025, EntitlementsV2025Api, EntitlementsV2025ApiListEntitlementsRequest, EntitlementV2025, PublicIdentitiesV2025Api, PublicIdentitiesV2025ApiGetPublicIdentitiesRequest, RoleV2025, TransformsV2025Api, TransformV2025, TransformReadV2025, SearchV2025Api, IndexV2025, AccessProfileDocumentV2025, EntitlementDocumentV2025, RoleDocumentV2025, EventDocumentV2025, AccountActivityDocumentV2025, AccessProfileV2025 } from 'sailpoint-api-client';
 import { DEFAULT_PUBLIC_IDENTITIES_QUERY_PARAMS } from '../models/PublicIdentity';
 import axios, { AxiosInstance, AxiosResponse } from 'axios';
 import { ImportEntitlementsResult } from '../models/JobStatus';
@@ -1632,40 +1632,80 @@ export class ISCClient {
 	//////////////////////////////
 
 
-	public async *getForms(filters: string | undefined = undefined): AsyncGenerator<FormBeta> {
+	public async *getForms(filters?: string): AsyncGenerator<FormDefinitionResponseV2025> {
 		console.log("> getForms");
 		const apiConfig = await this.getApiConfiguration();
-		const api = new CustomFormsBetaApi(apiConfig, undefined, this.getAxiosWithInterceptors());
-		let args = {
-			offset: 0,
-			limit: DEFAULT_PAGINATION,
-			filters
-		}
-		let count = -1
+		const api = new CustomFormsV2025Api(apiConfig, undefined, this.getAxiosWithInterceptors());
+		let offset = 0;
+		const limit = DEFAULT_PAGINATION;
+		let count = -1;
 		do {
-			const response = await api.searchFormDefinitionsByTenant(args)
-			count = response.data.count
+			const response = await api.searchFormDefinitionsByTenant({ offset, limit, filters });
+			count = response.data.count ?? 0;
 			if (response.data.results) {
 				for (const f of response.data.results) {
-					yield f
+					yield f;
 				}
 			}
-			args.offset += DEFAULT_PAGINATION
-
-			// if requesting an offset > text": "offset is greater than number of form definitions results"
-			// By using this criteria, we may fall on some edge cases where the total number of forms is a muliple of 250
-			// If using more than 200 Forms, the display would be slow, so it would mean that we would need to paginate like access profiles or roles.
-			// For now, we will stay like this
-		} while (count === DEFAULT_PAGINATION)
+			offset += limit;
+			// if requesting an offset > total: "offset is greater than number of form definitions results"
+			// By using this criteria, we may fall on some edge cases where the total is a multiple of 250.
+		} while (count === limit);
 	}
 
-	public async listForms(): Promise<FormDefinitionResponseBeta[]> {
+	public async listForms(): Promise<FormDefinitionResponseV2025[]> {
 		console.log("> listForms");
-		const forms: FormDefinitionResponseBeta[] = []
+		const forms: FormDefinitionResponseV2025[] = [];
 		for await (const form of this.getForms()) {
-			forms.push(form)
+			forms.push(form);
 		}
-		return forms
+		return forms;
+	}
+
+	public async getFormById(id: string): Promise<FormDefinitionResponseV2025> {
+		console.log("> getFormById", id);
+		const apiConfig = await this.getApiConfiguration();
+		const api = new CustomFormsV2025Api(apiConfig, undefined, this.getAxiosWithInterceptors());
+		const response = await api.getFormDefinitionByKey({ formDefinitionID: id });
+		return response.data;
+	}
+
+	public async getFormByName(name: string): Promise<FormDefinitionResponseV2025> {
+		console.log("> getFormByName", name);
+		const apiConfig = await this.getApiConfiguration();
+		const api = new CustomFormsV2025Api(apiConfig, undefined, this.getAxiosWithInterceptors());
+		const response = await api.searchFormDefinitionsByTenant({
+			filters: `name eq "${name}"`,
+			limit: 1,
+		});
+		const results = response.data.results;
+		if (!results || results.length === 0) {
+			throw new Error(`Form "${name}" not found.`);
+		}
+		return results[0];
+	}
+
+	public async createForm(payload: CreateFormDefinitionRequestV2025): Promise<FormDefinitionResponseV2025> {
+		console.log("> createForm");
+		const apiConfig = await this.getApiConfiguration();
+		const api = new CustomFormsV2025Api(apiConfig, undefined, this.getAxiosWithInterceptors());
+		const response = await api.createFormDefinition({ body: payload });
+		return response.data;
+	}
+
+	public async patchForm(id: string, patches: Array<{ [key: string]: object }>): Promise<FormDefinitionResponseV2025> {
+		console.log("> patchForm", id);
+		const apiConfig = await this.getApiConfiguration();
+		const api = new CustomFormsV2025Api(apiConfig, undefined, this.getAxiosWithInterceptors());
+		const response = await api.patchFormDefinition({ formDefinitionID: id, body: patches });
+		return response.data;
+	}
+
+	public async deleteFormById(id: string): Promise<void> {
+		console.log("> deleteFormById", id);
+		const apiConfig = await this.getApiConfiguration();
+		const api = new CustomFormsV2025Api(apiConfig, undefined, this.getAxiosWithInterceptors());
+		await api.deleteFormDefinition({ formDefinitionID: id });
 	}
 
 	public async exportForms(filters: string | undefined = undefined): Promise<ExportFormDefinitionsByTenant200ResponseInnerBeta[]> {